+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ v2.1.0 (IN PROGRESS)
 * Fix `notifications` returning nonsense when passing a single `id` (Thanks @chinchalinchin for the report)
 * Fix `moved` accidentally being named `moved_to_account` (Thanks @unusualevent for the report)
 * Added a warning for deprecated endpoints if the "deprecation" header is present
+* Added `oauth_userinfo` endpoint.
 
 v2.0.1
 ------